Items are now added to the database
Logged in users items are now stored in the database. No other calls are made to check if they are stored, so attempting to add the same item multiple times results in unseen errors.
Some additional minor changes were added to other pages including Purchase History which is no longer going to be used